Now UI reacts to clicks
It took me a while to understand all this dance on the UI state with
`remember` and `mutableStateOf()`. Probably a `ViewModel` would be
best, but I opted for a simpler solution.

All the composables needed to be under the same composable function
`GameBoard()` for that to work. Also, the hacky `boardRedrawCount`
must be shown on the screen (or used in a modifier) for the redraw
to be made.

Clicks are ignored for already matched cards.

Adopted the Scaffold layout, with a top bar. I still could not
make the colors work.