Cards appear turned down at start up
Also moved all the "clicked" logic to the model from the UI handler.

Improved when exactly we should react to a click. We're ignoring
already matched and cards with face up in the first move (to avoid
undos).

Introduced a new state: MOVE3. This is the next click when two
cards were turned but there's no match. This move will just reset
the board turning both cards down.